update met jelles dingen

+++ videoPagina.php
@@ -5,10 +5,18 @@ require_once("./security.php");
 
 <?php
 if (isset($_POST['reserveer'])) {
-    echo "<h3 style='text-align: center;' >Item toegevoegd aan reserveringen.</h3><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br><br>";
-    header("refresh:4;url=index.php?content=reserveringen");
     require_once("./classes/ReserveClass.php");
-    ReserveClass::insert_reserveringitem_database($_POST);
+    if (!ReserveClass::check_if_reservering_exists($_POST))
+    {
+        echo "<h3 style='text-align: center;' >Item toegevoegd aan reserveringen.</h3><br><br><br><br><br><br><br><br>           <br><br><br><br><br><br><br><br><br><br><br><br><br><br><br>";
+        header("refresh:4;url=index.php?content=reserveringen");
+        
+        ReserveClass::insert_reserveringitem_database($_POST);
+    }
+    else {
+        
+        echo "U heeft deze video al gereserveerd";
+    }
 } else {
     if (isset($_POST['submit'])) {
 
@@ -59,6 +67,12 @@ if (isset($_POST['reserveer'])) {
                     if ($conn->connect_error) {
                         die("Connection failed: " . $conn->connect_error);
                     }
+                    if ($conn2->connect_error) {
+                        die("Connection failed: " . $conn2->connect_error);
+                    }
+                    if ($conn3->connect_error) {
+                        die("Connection failed: " . $conn3->connect_error);
+                    }
 
                     $idVideo = $_GET['idVideo'];
                     $sql = "SELECT * FROM `video` 
@@ -95,7 +109,7 @@ if (isset($_POST['reserveer'])) {
                             if ($row["aantalBeschikbaar"] > 0) {
                                 echo "<b>Aantal beschikbaar: </b>" . $row["aantalBeschikbaar"] . "<br><br>";
                             } else {
-                                echo "<b>Deze video is helaas uitverkocht plaats een reservering om de video te kunnen huren als die weer beschikbaar is</b>";
+                                echo "<b>Deze video is helaas uitverkocht. Plaats een reservering om de video te kunnen huren als die weer beschikbaar is</b>";
                             }
                             echo "
                             <b>Prijs: </b>
